Updated Coordinates & AddressCoordinatesDTO to use double instead of Big double for easier compat w Redis

+++ API/src/main/java/com/citycite_api/report/dto/AddressCoordinatesDTO.java
@@ -1,12 +1,11 @@
 package com.citycite_api.report.dto;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class AddressCoordinatesDTO {
 
-    private BigDecimal longitude;
-    private BigDecimal latitude;
+    private double longitude;
+    private double latitude;
 
 }
